scripts/requirements.txt
psycopg2-binary>=2.9.3
python-dotenv>=0.19.2
pandas>=1.3.3
numpy>=1.21.2
Faker>=13.3.0
